Fix stof return type error and other windows warnings (#2254)
Submitted on behalf of Patrick Hodoul, with some minor mods by LG.

* Fixes error in return type of stof.

* Add several casts that fix MSVS warnings.

+++ src/include/OpenImageIO/ustring.h
@@ -850,7 +850,7 @@ iequals(const std::string& a, ustring b)
 // ustring variant stof from OpenImageIO/strutil.h
 namespace Strutil {
 
-inline int
+inline float
 stof(ustring s)
 {
     return Strutil::stof(s.string());
